Crate mirrord_layer

Loaded dynamically with your local process.

Paired with [mirrord-agent], it makes your local process behave as if it was running in a remote context.

Check out the Introduction guide to learn more about mirrord.

How it works

This crate intercepts your processes’ libc calls, and instead of executing them locally (as normal), it instead forwards them as a message to the mirrord-agent pod. The operation is executed there, with the result being returned back to mirrord-layer, and finally to the original libc call.

Example

Let’s say you have a Node.js app that just opens a file, like this:

open-file.mjs

import { open } from 'node:fs';

const file = open('/tmp/hello.txt');

When run with mirrord, this is what’s going to happen:

We intercept the call to libc::open using our open_detour hook, which calls file::ops::open;

file::ops::open sends an open file message to mirrord-agent;

mirrore-agent tries to open /tmp/hello.txt in the remote context it’s running, and returns the result of the operation back to mirrord-layer;

We handle the mapping of the remote file (the one we have open in mirrord-agent), and a local file (temporarily created);

And finally, we return the expected result (type) to your Node.js application, as if it had just called libc::open.

Your application will get an fd that is valid in the context of mirrord, and calls to other file functions (like libc::read), will work just fine, operating on the remote file.

Configuration

The functions we intercept are controlled via the mirrord-config crate, check its documentation for more details, or Configuration for usage information.

Modules

common 🔒

Shared place for a few types and functions that are used everywhere by the layer.

connection 🔒

Module for the mirrord-layer/mirrord-agent connection mechanism.

detour 🔒

The layer uses features from this module to check if it should bypass one of its hooks, and call the original libc function.

dns 🔒

Holds the GetAddrInfo HookMessage used to perform DNS lookup in a remote context.

error 🔒

file 🔒

go_hooks 🔒

hooks 🔒

load 🔒

macros 🔒

Macros used by mirrord-layer (mostly for hooking libc functions).

outgoing 🔒

socket 🔒

We implement each hook function in a safe function as much as possible, having the unsafe do the absolute minimum

tcp 🔒

tcp_mirror 🔒

tcp_steal 🔒

tracing_util 🔒

Macros

graceful_exit

Kills the process and prints a helpful error message to the user.

replace

Replaces the $func libc function, with the equivalent hook $detour_function, by calling HookManager::hook_export_or_any.

Structs

Layer 🔒

Acts as an API to the various features of mirrord-layer, holding the actual feature handler types.

Constants

BRANCHES 🔒

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

ENABLED_TCP_OUTGOING 🔒

Tells us if the user enabled the Tcp outgoing feature in [OutgoingConfig].

ENABLED_UDP_OUTGOING 🔒

Tells us if the user enabled the Udp outgoing feature in [OutgoingConfig].

FILE_MODE 🔒

Holds the file operations configuration, as specified by FsConfig.

FN_CLOSE 🔒

FN_CLOSE_NOCANCEL 🔒

FN_UV_FS_CLOSE 🔒

FN___CLOSE 🔒

FN___CLOSE_NOCANCEL 🔒

HOOK_SENDER 🔒

Sender for the HookMessages that are handled internally, and converted (when applicable) to ClientMessages.

INCOMING_IGNORE_LOCALHOST 🔒

Tells us if the user enabled wants to ignore listening on localhost in IncomingConfig.

INCOMING_IGNORE_PORTS 🔒

Ports to ignore on listening for mirroring/stealing.

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

OUTGOING_IGNORE_LOCALHOST 🔒

Tells us if the user enabled wants to ignore localhots connections in [OutgoingConfig].

REMOTE_UNIX_STREAMS 🔒

Unix streams to connect to remotely.

RUNTIME 🔒

Main tokio Runtime for mirrord-layer async tasks.

Functions

__close_detour 🔒 ⚠

__close_nocancel_detour 🔒 ⚠

close_detour 🔒 ⚠

Attempts to close on a managed Socket, if there is no socket with fd, then this means we either let the fd bypass and call libc::close directly, or it might be a managed file fd, so it tries to do the same for files.

close_layer_fd 🔒

Shared code for closing fd in our data structures.

close_nocancel_detour 🔒 ⚠

No need to guard because we call another detour which will do the guard for us.

enable_hooks 🔒

Prepares the HookManager and replace!s libc calls with our hooks, according to what the user configured.

is_nix_or_devbox 🔒

Check if NixOS or Devbox by discrimnating env vars.

layer_pre_initialization 🔒

Loads mirrord configuration and applies nix_devbox_patch patches.

layer_start 🔒

Occurs after layer_pre_initialization has succeeded.

mirrord_layer_entry_point 🔒

The one true start of mirrord-layer.

nix_devbox_patch 🔒

Check if we’re running in NixOS or Devbox.

port_debug_patch 🔒

Prevent mirrord from connecting to ports used by the intelliJ debugger

start_layer_thread 🔒

Initializes mirrord-layer’s thread_loop in a separate tokio::task.

thread_loop 🔒

Main loop of mirrord-layer.

uv_fs_close_detour 🔒 ⚠

Hook

Type Definitions

FnClose 🔒

FnClose_nocancel 🔒

FnUv_fs_close 🔒

Fn__close 🔒

Fn__close_nocancel 🔒

Opaque Types

🔒

🔒

🔒

🔒

🔒