Function mirrord_layer::connection::connect

pub(crate) async fn connect(
    config: &LayerConfig
) -> (Sender<ClientMessage>, Receiver<DaemonMessage>)

Initializes the agent pod with KubernetesAPI, and connects to it, returning an API that the layer uses to communicate with it, in the form of a Sender for ClientMessages, and a Receiver for DaemonMessages.

Direct connection

There is support for connecting directly to an existing pod (instead of going through KubernetesAPI::create) by setting LayerConfig::connect_tcp.

This is used for mirrord tests only.

Operator

The OperatorApi takes over the connection procedure if LayerConfig::operator is set to true and an operator is installed in the cluster. We use OperatorApi::discover instead of KubernetesAPI.