Module mirrord_layer::detour

The layer uses features from this module to check if it should bypass one of its hooks, and call the original libc function.

Here we also have the convenient Detour, that is used by the hooks to either return a Result-like value, or the special Bypass case, which makes the detour function call the original libc equivalent, stored in a HookFn.

Structs

DetourGuard 🔒

Handler for the layer’s DETOUR_BYPASS.

HookFn 🔒

Wrapper around OnceLock, mainly used for the Deref implementation to simplify calls to the original functions as FN_ORIGINAL(), instead of FN_ORIGINAL.get().unwrap().

Enums

Bypass 🔒

Soft-errors that can be recovered from by calling the raw FFI function.

Detour 🔒

ControlFlow-like enum to be used by hooks.

Constants

DETOUR_BYPASS 🔒

Holds the thread-local state for bypassing the layer’s detour functions.

Traits

OnceLockExt 🔒

Extends OnceLock with a helper function to initialize it with a Detour.

OptionExt 🔒

Extends Option<T> with the Option::bypass function.

Functions

detour_bypass_off 🔒

Sets DETOUR_BYPASS to false.

detour_bypass_on 🔒

Sets DETOUR_BYPASS to true, bypassing the layer’s detours.