Module mirrord_layer::file::filter

Macros

some_regex_match 🔒

Shortcut for checking an optional regex set. Usage is some_regex_match(Option<RegexSet>, text)

Structs

FileFilter 🔒

Statics

CALLSITE 🔒

FILE_FILTER 🔒

Global filter used by file operations to bypass (use local) or continue (use remote).

Functions

build_regex_or_none 🔒

Builds case insensitive regexes from a list of patterns. Returns None if the list is empty. Regex error if fails.

generate_local_set 🔒

List of files that mirrord should use locally, as they probably exist only in the local user machine, or are system configuration files (that could break the process if we used the remote version).

generate_remote_ro_set 🔒

List of files that mirrord should use remotely read only