Function mirrord_layer::layer_start

pub(crate) fn layer_start(config: LayerConfig)

Occurs after layer_pre_initialization has succeeded.

Starts the main parts of mirrord-layer.

Details

Sets up a few things based on the LayerConfig given by the user:

tracing_subscriber, or mirrord_console;

Connects to the mirrord-agent with [connection::connect];

Initializes some of our globals;

Replaces the libc calls with our hooks with enable_hooks;

Starts the main mirrord-layer thread.