Function mirrord_layer::thread_loop

pub(crate) async fn thread_loop(
    receiver: Receiver<HookMessage>,
    tx: Sender<ClientMessage>,
    rx: Receiver<DaemonMessage>,
    config: LayerConfig
)

Main loop of mirrord-layer.

Parameters

receiver: receives the HookMessages and pass them along with Layer::handle_hook_message;

tx, rx: see Layer::tx](link), and [Layer::rx;

config: the mirrord-layer configuration, see LayerConfig.

Details

We have our big loop here that is initialized in a separate task by start_layer_thread.

In this loop we:

Pass HookMessages to be handled by the Layer;

Read from the Layer feature handlers Receivers, to turn outgoing messages into ClientMessages, sending them with Layer::send;

Handle the heartbeat mechanism (Ping/Pong feature), sending a ClientMessage::Ping if all the other channels received nothing for a while (60 seconds);

Write log file if FeatureConfig::capture_error_trace is set.