Module mirrord_layer::socket

We implement each hook function in a safe function as much as possible, having the unsafe do the absolute minimum

Modules

hooks ğŸ”’

id ğŸ”’

The types in this module should NOT be constructed directly.

ops ğŸ”’

Structs

Bound

Represents a SocketState where the user made a libc::bind call, and we intercepted it.

Connected

Contains the addresses of a mirrord connected socket.

ConnectionQueue

poll_agent loop inserts connection data into this queue, and accept reads it.

SocketInformation ğŸ”’

Struct sent over the socket once created to pass metadata to the hook

UserSocket ğŸ”’

Enums

SocketKind ğŸ”’

SocketState

Statics

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CONNECTION_QUEUE

Holds the connections that have yet to be accepted.

SOCKETS ğŸ”’

Traits

ProtocolExt ğŸ”’

SocketAddrExt ğŸ”’

Trait that expands std and socket2 sockets.

Functions

fill_address ğŸ”’

Fill in the sockaddr structure for the given address.

is_ignored_port ğŸ”’