Trait mirrord_layer::tcp_steal::http::HttpV

pub(super) trait HttpV: Sized {
    type Sender;
    type Connection: Future + Send + 'static;

    // Required methods
    fn new(http_request_sender: Self::Sender) -> Self;
    async fn handshake(
        target_stream: TcpStream
    ) -> Result<(Self::Sender, Self::Connection), HttpForwarderError>;
    fn take_sender(self) -> Self::Sender;
    fn set_sender(&mut self, new_sender: Self::Sender);
    async fn send_request(
        &mut self,
        request: HttpRequest
    ) -> Result<Response<Incoming>>;

    // Provided method
    async fn send_http_request_to_application(
        &mut self,
        request: HttpRequest,
        port: Port,
        connection_id: ConnectionId,
        destination: SocketAddr
    ) -> Result<HttpResponse, HttpForwarderError> { ... }
}

Handles the differences between hyper’s HTTP/1 and HTTP/2 connections.

Thanks to this trait being implemented for both v1::HttpV1 and v2::HttpV2, we can have most of the implementation for dealing with requests in ConnectionTask.

Required Associated Types

type Sender

type Connection: Future + Send + 'static

Required Methods

fn new(http_request_sender: Self::Sender) -> Self

Creates a new ConnectionTask that handles [HttpV1]/[HttpV2] requests.

Connects to the user’s application with Self::connect_to_application.

async fn handshake( target_stream: TcpStream ) -> Result<(Self::Sender, Self::Connection), HttpForwarderError>

Calls the appropriate HTTP/V handshake method.

fn take_sender(self) -> Self::Sender

Takes the HttpV::Sender, inserting None back into this instance.

Usage

Mainly used for our retry connection mechanism, where we create a new HttpV implementor, but we want to swap the “failed attempt” HttpV::Sender with the newly created one.

fn set_sender(&mut self, new_sender: Self::Sender)

Sets this instance’s HttpV::Sender.

Usage

Required as part of the HttpV::take_sender retry mechanism.

async fn send_request( &mut self, request: HttpRequest ) -> Result<Response<Incoming>>

Sends the HttpRequest to its destination.

Required due to this mechanism having distinct types; HTTP/1 with http1::SendRequest, and HTTP/2 with http2::SendRequest.

Provided Methods

async fn send_http_request_to_application( &mut self, request: HttpRequest, port: Port, connection_id: ConnectionId, destination: SocketAddr ) -> Result<HttpResponse, HttpForwarderError>

Sends the HttpRequest through Self::sender, converting the response into a HttpResponse.