Trait mirrord_layer::common::CheckedInto

pub(crate) trait CheckedInto<T>: Sized {
    fn checked_into(self) -> Detour<T>;
}

Converts raw pointer values P to some other type.

Usage

Mainly used to convert from raw C strings (*const c_char) into a Rust type wrapped in Detour.

These conversions happen in the unsafe hook functions, and we pass the converted value inside a Detour to defer the handling of null pointers (and other invalid-ish values) when the ops version of the function returns an Error or Bypass.

Required Methods

fn checked_into(self) -> Detour<T>

Converts Self to Detour<T>.