Module mirrord_layer::connection

Module for the mirrord-layer/mirrord-agent connection mechanism.

The layer will either start a new agent pod with KubernetesAPI, directly connect to an existing agent (currently only used for tests), or let the OperatorApi handle the connection.

Constants

FAIL_CREATE_AGENT 🔒

Helpful message that we print to the user in a graceful_exit call, when we get a KubeApiError, or an `OperatorApiError.

Functions

connect 🔒

Initializes the agent pod with KubernetesAPI, and connects to it, returning an API that the layer uses to communicate with it, in the form of a Sender for ClientMessages, and a Receiver for DaemonMessages.

handle_error 🔒

Handles a KubeApiError by printing what went wrong to the user, and closing the program with graceful_exit.

handle_operator_error 🔒

Calls graceful_exit when we have an OperatorApiError.