Module mirrord_layer::connection

Module for the mirrord-layer/mirrord-agent connection mechanism.

The layer will connect to the internal proxy to communicate with the agent.

Constants

CONNECTION_CHANNEL_SIZE 🔒

Size of the channel used in wrap_raw_connection for ClientMessages.

FAIL_STILL_STUCK 🔒

Functions

connect_to_proxy 🔒

Connects to the internal proxy in given SocketAddr layer uses to communicate with it, in the form of a Sender for ClientMessages, and a Receiver for DaemonMessages.

wrap_raw_connection 🔒

Creates the task that handles the messaging between layer/agent. It does the encoding/decoding of protocol.