Module mirrord_layer::debugger_ports

Enums

DebuggerPorts

Local ports used by the debugger running the process. These should be ignored by the layer.

DebuggerType

Type of debugger which is used to run the user’s processes. Determines the way we parse the command line arguments the debugger’s port. Logic of processing the arguments is based on examples taken from the IDEs.

Constants

MIRRORD_DETECT_DEBUGGER_PORT_ENV

Environment variable used to tell the layer that it should dynamically detect the local port used by the given debugger. Value passed through this variable should parse into DebuggerType. Used when injecting the layer through IDE, because the debugger port is chosen dynamically.

MIRRORD_IGNORE_DEBUGGER_PORTS_ENV

Environment variable used to tell the layer that it should ignore certain local ports that may be used by the debugger. Used when injecting the layer through IDE. This setting will be ignored if the layer successfully detects the port at runtime, see MIRRORD_DETECT_DEBUGGER_PORT_ENV.

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

META 🔒

META 🔒

META 🔒

META 🔒