Module mirrord_layer::file::ops

Macros

should_ignore ğŸ”’

should_ignore(path: P, write: bool) Checks if the file should be ignored, and returns a Bypass::IgnoredFile if it should.

Structs

RemoteFile ğŸ”’

Functions

access ğŸ”’

blocking_send_file_message ğŸ”’

close_remote_file_on_failure ğŸ”’

Close the remote file if the call to libc::shm_open failed and we have an invalid local fd.

closedir ğŸ”’

create_local_fake_file ğŸ”’ âš

The pair libc::shm_open, libc::shm_unlink are used to create a temporary file (in /dev/shm/), and then remove it, as we only care about the fd. This is done to preserve open_flags, as libc::memfd_create will always return a File with read and write permissions (which is undesirable).

fdopen ğŸ”’

fdopendir ğŸ”’

creates a directory stream for the remote_fd in the agent

fgets ğŸ”’

fopen ğŸ”’

Calls open and returns a FILE pointer based on the local fd.

get_remote_fd ğŸ”’

Helper function that retrieves the remote_fd (which is generated by mirrord_agent::util::IndexAllocator).

getdents64 ğŸ”’

lseek ğŸ”’

open ğŸ”’

Blocking wrapper around libc::open call.

openat ğŸ”’

path_from_rawish ğŸ”’

Converts a CStr path into a PathBuf, or bypasses when this fails.

pread ğŸ”’

pwrite ğŸ”’

read ğŸ”’

Blocking wrapper around libc::read call.

readdir_r ğŸ”’

str_from_rawish ğŸ”’

Converts a CStr path into a &str, or bypasses when this fails.

write ğŸ”’

xstat ğŸ”’

General stat function that can be used for lstat, fstat, stat and fstatat. Note: We treat cases of AT_SYMLINK_NOFOLLOW_ANY as AT_SYMLINK_NOFOLLOW because even Go does that. rawish_path is Option<Option<&CStr>> because we need to differentiate between null pointer and non existing argument (For error handling)