Module mirrord_layer::file::ops

Macros

should_ignore ğŸ”’

should_ignore(path: P, write: bool) Checks if the file should be ignored, and returns a Bypass::IgnoredFile if it should.

Structs

RemoteFile ğŸ”’

Constants

MAX_READ_SIZE ğŸ”’

1 Megabyte. Large read requests can lead to timeouts.

Statics

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

Functions

access ğŸ”’

blocking_send_file_message ğŸ”’

close_remote_file_on_failure ğŸ”’

Close the remote file if the call to libc::shm_open failed and we have an invalid local fd.

closedir ğŸ”’

create_local_fake_file ğŸ”’

Create temporary local file to get a valid local fd.

fdopen ğŸ”’

fdopendir ğŸ”’

creates a directory stream for the remote_fd in the agent

get_remote_fd ğŸ”’

Helper function that retrieves the remote_fd (which is generated by mirrord_agent::util::IndexAllocator).

getdents64 ğŸ”’

lseek ğŸ”’

open ğŸ”’

Blocking wrapper around libc::open call.

openat ğŸ”’

pread ğŸ”’

pwrite ğŸ”’

read ğŸ”’

Blocking wrapper around libc::read call.

readdir_r ğŸ”’

write ğŸ”’

xstat ğŸ”’

General stat function that can be used for lstat, fstat, stat and fstatat. Note: We treat cases of AT_SYMLINK_NOFOLLOW_ANY as AT_SYMLINK_NOFOLLOW because even Go does that. rawish_path is Option<Option<&CStr>> because we need to differentiate between null pointer and non existing argument (For error handling)

xstatfs ğŸ”’