Trait mirrord_layer::tcp::TcpHandler

pub(crate) trait TcpHandler {
    fn ports(&self) -> &HashSet<Listen>;
    fn ports_mut(&mut self) -> &mut HashSet<Listen>;
    fn handle_new_connection<'life0, 'async_trait>(
        &'life0 mut self,
        conn: NewTcpConnection
    ) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn handle_new_data<'life0, 'async_trait>(
        &'life0 mut self,
        data: TcpData
    ) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn handle_http_request<'life0, 'async_trait>(
        &'life0 mut self,
        request: HttpRequest
    ) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn handle_close(&mut self, close: TcpClose) -> Result<(), LayerError>;
    fn handle_listen<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        listen: Listen,
        tx: &'life1 Sender<ClientMessage>
    ) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;

    fn handle_daemon_message<'life0, 'async_trait>(
        &'life0 mut self,
        message: DaemonTcp
    ) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>
    where
        Self: Send + 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn handle_hook_message<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        message: TcpIncoming,
        tx: &'life1 Sender<ClientMessage>
    ) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>
    where
        Self: Send + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
    { ... }
    fn create_local_stream<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        tcp_connection: &'life1 NewTcpConnection
    ) -> Pin<Box<dyn Future<Output = Result<TcpStream, LayerError>> + Send + 'async_trait>>
    where
        Self: Send + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
    { ... }
}

Required Methods

fn ports(&self) -> &HashSet<Listen>

fn ports_mut(&mut self) -> &mut HashSet<Listen>

fn handle_new_connection<'life0, 'async_trait>(
&'life0 mut self,
conn: NewTcpConnection
) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Handle NewConnection messages

fn handle_new_data<'life0, 'async_trait>(
&'life0 mut self,
data: TcpData
) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Handle New Data messages

fn handle_http_request<'life0, 'async_trait>(
&'life0 mut self,
request: HttpRequest
) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Handle New Data messages

fn handle_close(&mut self, close: TcpClose) -> Result<(), LayerError>

Handle connection close

fn handle_listen<'life0, 'life1, 'async_trait>(
&'life0 mut self,
listen: Listen,
tx: &'life1 Sender<ClientMessage>
) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Handle listen request

Provided Methods

fn handle_daemon_message<'life0, 'async_trait>(
&'life0 mut self,
message: DaemonTcp
) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>where
Self: Send + 'async_trait,
'life0: 'async_trait,

Returns true to let caller know to keep running

fn handle_hook_message<'life0, 'life1, 'async_trait>(
&'life0 mut self,
message: TcpIncoming,
tx: &'life1 Sender<ClientMessage>
) -> Pin<Box<dyn Future<Output = Result<(), LayerError>> + Send + 'async_trait>>where
Self: Send + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn create_local_stream<'life0, 'life1, 'async_trait>(
&'life0 mut self,
tcp_connection: &'life1 NewTcpConnection
) -> Pin<Box<dyn Future<Output = Result<TcpStream, LayerError>> + Send + 'async_trait>>where
Self: Send + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Connects to the local listening socket, add it to the queue and return the stream. Find better name