Trait mirrord_layer::tcp::TcpHandler

pub(crate) trait TcpHandler {
    // Required methods
    fn ports(&self) -> &HashSet<Listen>;
    fn ports_mut(&mut self) -> &mut HashSet<Listen>;
    fn port_mapping_ref(&self) -> &BiMap<u16, u16>;
    async fn handle_new_connection(
        &mut self,
        conn: NewTcpConnection
    ) -> Result<(), LayerError>;
    async fn handle_new_data(&mut self, data: TcpData) -> Result<(), LayerError>;
    async fn handle_http_request(
        &mut self,
        request: HttpRequest
    ) -> Result<(), LayerError>;
    fn handle_close(&mut self, close: TcpClose) -> Result<(), LayerError>;
    async fn handle_listen(
        &mut self,
        listen: Listen,
        tx: &Sender<ClientMessage>
    ) -> Result<(), LayerError>;

    // Provided methods
    fn apply_port_mapping(&self, listen: &mut Listen) { ... }
    async fn handle_daemon_message(
        &mut self,
        message: DaemonTcp
    ) -> Result<(), LayerError> { ... }
    async fn handle_hook_message(
        &mut self,
        message: TcpIncoming,
        tx: &Sender<ClientMessage>
    ) -> Result<(), LayerError> { ... }
    async fn create_local_stream(
        &mut self,
        tcp_connection: &NewTcpConnection
    ) -> Result<TcpStream, LayerError> { ... }
}

Required Methods

fn ports(&self) -> &HashSet<Listen>

fn ports_mut(&mut self) -> &mut HashSet<Listen>

fn port_mapping_ref(&self) -> &BiMap<u16, u16>

async fn handle_new_connection( &mut self, conn: NewTcpConnection ) -> Result<(), LayerError>

Handle NewConnection messages

async fn handle_new_data(&mut self, data: TcpData) -> Result<(), LayerError>

Handle New Data messages

async fn handle_http_request( &mut self, request: HttpRequest ) -> Result<(), LayerError>

Handle New Data messages

fn handle_close(&mut self, close: TcpClose) -> Result<(), LayerError>

Handle connection close

async fn handle_listen( &mut self, listen: Listen, tx: &Sender<ClientMessage> ) -> Result<(), LayerError>

Handle listen request

Provided Methods

fn apply_port_mapping(&self, listen: &mut Listen)

Modify Listen to match local port to remote port based on mapping If no mapping is found, the port is not modified

async fn handle_daemon_message( &mut self, message: DaemonTcp ) -> Result<(), LayerError>

Returns true to let caller know to keep running

async fn handle_hook_message( &mut self, message: TcpIncoming, tx: &Sender<ClientMessage> ) -> Result<(), LayerError>

async fn create_local_stream( &mut self, tcp_connection: &NewTcpConnection ) -> Result<TcpStream, LayerError>

Connects to the local listening socket, add it to the queue and return the stream. Find better name